Cleaned old tests
